refactor: 단일책임원칙에 따른 분리
