--- src/main/java/org/sopt/carena/member/appliacation/service/JwtTokenProvider.java
@@ -1,77 +1,0 @@
-package org.sopt.carena.member.appliacation.service;
-
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.JwtException;
-import io.jsonwebtoken.SignatureAlgorithm;
-import io.jsonwebtoken.security.Keys;
-import lombok.extern.slf4j.Slf4j;
-import org.sopt.carena.member.appliacation.exception.jwt.InvalidTokenException;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-import java.security.Key;
-import java.util.Date;
-
-
-@Slf4j
-@Component
-public class JwtTokenProvider {
-
-    private final Key key;
-    private final long accessTokenValidityInMilliseconds;
-    private final long refreshTokenValidityInMilliseconds;
-
-    public JwtTokenProvider(
-            @Value("${jwt.secret}") String secret,
-            @Value("${jwt.access-token-expiration}") long accessTokenValidityInMilliseconds,
-            @Value("${jwt.refresh-token-expiration}") long refreshTokenValidityInMilliseconds
-    ) {
-        this.key = Keys.hmacShaKeyFor(secret.getBytes());
-        this.accessTokenValidityInMilliseconds = accessTokenValidityInMilliseconds;
-        this.refreshTokenValidityInMilliseconds = refreshTokenValidityInMilliseconds;
-    }
-
-    public String createAccessToken(Long memberId) {
-        return createToken(memberId, accessTokenValidityInMilliseconds);
-    }
-
-    public String createRefreshToken(Long memberId) {
-        return createToken(memberId, refreshTokenValidityInMilliseconds);
-    }
-
-    public String createToken(Long memberId,long validityInMilliseconds) {
-        Date now = new Date();
-        Date validity = new Date(now.getTime() + validityInMilliseconds);
-
-        return Jwts.builder()
-                .setSubject(String.valueOf(memberId))
-                .setIssuedAt(now)
-                .setExpiration(validity)
-                .signWith(key, SignatureAlgorithm.HS256)
-                .compact();
-    }
-
-    public boolean validateToken(String token) {
-        try {
-            Jwts.parser()
-                    .setSigningKey(key)
-                    .build()
-                    .parseClaimsJws(token);
-            return true;
-        } catch (JwtException | IllegalArgumentException e) {
-            log.error("JWT 검증 실패", e);
-            throw new InvalidTokenException();
-        }
-    }
-
-    public Long getMemberIdFromToken(String token) {
-        Claims claims = Jwts.parser()
-                .setSigningKey(key)
-                .build()
-                .parseClaimsJws(token)
-                .getBody();
-
-        return Long.parseLong(claims.getSubject());
-    }
-}
