+++ src/main/java/org/sopt/carena/member/appliacation/service/util/JwtProperties.java
@@ -1,0 +1,38 @@
+package org.sopt.carena.member.appliacation.service.util;
+
+import io.jsonwebtoken.security.Keys;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+import java.security.Key;
+
+// JWT 설정 관리
+@Component
+public class JwtProperties {
+
+    private final Key key;
+    private final long accessTokenValidityInMilliseconds;
+    private final long refreshTokenValidityInMilliseconds;
+
+    public JwtProperties(
+            @Value("${jwt.secret}") String secret,
+            @Value("${jwt.access-token-expiration}") long accessTokenValidityInMilliseconds,
+            @Value("${jwt.refresh-token-expiration}") long refreshTokenValidityInMilliseconds
+    ) {
+        this.key = Keys.hmacShaKeyFor(secret.getBytes());
+        this.accessTokenValidityInMilliseconds = accessTokenValidityInMilliseconds;
+        this.refreshTokenValidityInMilliseconds = refreshTokenValidityInMilliseconds;
+    }
+
+    public Key getKey() {
+        return key;
+    }
+
+    public long getAccessTokenValidityInMilliseconds() {
+        return accessTokenValidityInMilliseconds;
+    }
+
+    public long getRefreshTokenValidityInMilliseconds() {
+        return refreshTokenValidityInMilliseconds;
+    }
+}
