+++ src/main/java/org/sopt/carena/member/appliacation/service/util/JwtTokenParser.java
@@ -1,0 +1,27 @@
+package org.sopt.carena.member.appliacation.service.util;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+//  토큰 파싱 책임
+@Component
+@RequiredArgsConstructor
+public class JwtTokenParser {
+
+    private final JwtProperties jwtProperties;
+
+    public Long getMemberId(String token) {
+        Claims claims = getClaims(token);
+        return Long.parseLong(claims.getSubject());
+    }
+
+    private Claims getClaims(String token) {
+        return Jwts.parser()
+                .setSigningKey(jwtProperties.getKey())
+                .build()
+                .parseClaimsJws(token)
+                .getBody();
+    }
+}
