+++ src/main/java/org/sopt/carena/member/appliacation/service/util/JwtTokenValidator.java
@@ -1,0 +1,30 @@
+package org.sopt.carena.member.appliacation.service.util;
+
+import io.jsonwebtoken.JwtException;
+import io.jsonwebtoken.Jwts;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.sopt.carena.member.exception.jwt.InvalidTokenException;
+import org.springframework.stereotype.Component;
+
+// 토큰 검증 책임
+@Slf4j
+@Component
+@RequiredArgsConstructor
+public class JwtTokenValidator {
+
+    private final JwtProperties jwtProperties;
+
+    public boolean validateToken(String token) {
+        try {
+            Jwts.parser()
+                    .setSigningKey(jwtProperties.getKey())
+                    .build()
+                    .parseClaimsJws(token);
+            return true;
+        } catch (JwtException | IllegalArgumentException e) {
+            log.error("JWT 검증 실패", e);
+            throw new InvalidTokenException();
+        }
+    }
+}
